Minimally start moving perf-specific code into a perf module.
The "collect" module hard-coded all sorts of perf things directly into
it. This commit starts the journey of separating out "perf specific
things" from the collect module. This is far from complete, but is a
minimal "compiles and passes all tests".
