+++ hwtracer/src/perf/mod.rs
@@ -1,0 +1,36 @@
+use libc::{size_t, sysconf, _SC_PAGESIZE};
+use std::{convert::TryFrom, sync::LazyLock};
+
+const PERF_DFLT_DATA_BUFSIZE: size_t = 64;
+static PERF_DFLT_AUX_BUFSIZE: LazyLock<size_t> = LazyLock::new(|| {
+    // Allocate enough pages for a 64MiB trace buffer.
+    let mb64 = 1024 * 1024 * 64;
+    let page_sz = size_t::try_from(unsafe { sysconf(_SC_PAGESIZE) }).unwrap();
+    mb64 / page_sz + size_t::from(mb64 % page_sz != 0)
+});
+
+const PERF_DFLT_INITIAL_TRACE_BUFSIZE: size_t = 1024 * 1024; // 1MiB
+
+/// Configures the Perf collector.
+///
+// Must stay in sync with the C code.
+#[derive(Clone, Debug)]
+#[repr(C)]
+pub struct PerfCollectorConfig {
+    /// Data buffer size, in pages. Must be a power of 2.
+    pub data_bufsize: size_t,
+    /// AUX buffer size, in pages. Must be a power of 2.
+    pub aux_bufsize: size_t,
+    /// The initial trace storage buffer size (in bytes) of new traces.
+    pub initial_trace_bufsize: size_t,
+}
+
+impl Default for PerfCollectorConfig {
+    fn default() -> Self {
+        Self {
+            data_bufsize: PERF_DFLT_DATA_BUFSIZE,
+            aux_bufsize: *PERF_DFLT_AUX_BUFSIZE,
+            initial_trace_bufsize: PERF_DFLT_INITIAL_TRACE_BUFSIZE,
+        }
+    }
+}
